reword confusing use of 'referring to'
I assume that the original intent was something like
"the expression 'triple term' refers to an RDF triple used as the object of another triple"
but the sentence is about the *construct*, not the *expression*.
It may be read as "a triple term refers to/denotes an RDF triple", which is wrong.

I think the proposed rewording respects the original intent,
but removes the confusion.